Switch to system notifications, add to feature preview, add error handling
